patientActivities/requirements.txt
boto3==1.20.42
uvicorn
fastapi
python-dotenv